% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_MMV.R
\name{multi_matrix}
\alias{multi_matrix}
\title{multiply two matrix}
\usage{
multi_matrix(M1, M2)
}
\arguments{
\item{M1}{a matrix}

\item{M2}{a matrix}
}
\value{
Multiply M1, M2
}
\description{
multiply two matrix
}
\examples{
M1=matrix(c(1,2,3,9),ncol=2)
M2=matrix(c(9,2,3,9),ncol=2)
multi_matrix(M1,M2)
}
